requirements.txt
paho-mqtt
influxdb
requests
colorama
